26th commit from Laptop

+++ Taxes/src/TaxType/TaxType.java
@@ -3,7 +3,7 @@ package TaxType;
 import java.math.BigDecimal;
 
 public class TaxType {
-    public long calculateTaxFor(BigDecimal amount) {
+    public BigDecimal calculateTaxFor(BigDecimal amount) {
         // TODO override me!
         return BigDecimal.valueOf(0.0);
     }
